requirements.txt
# ═══════════════════════════════════════════════════════════
# pip install -r requirements.txt
# ═══════════════════════════════════════════════════════════

python-dotenv>=1.0
requests>=2.31
beautifulsoup4>=4.12
playwright>=1.40
groq>=0.4
pyyaml>=6.0
Pillow>=10.0
google-api-python-client>=2.140
google-auth-oauthlib>=1.2
google-auth-httplib2>=0.1
flask>=3.0

